Acesso
Personalização dos botões conforme o nível de acesso.

+++ adm/app/adms/Views/accessLevels/listAcessLevels.php
@@ -11,8 +11,12 @@
             <span class="title-content">Niveis de Acesso</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" .URLADM . "add-access-levels/index' class='btn-success'>Cadastrar</a> ";
-                echo "<a href='" .URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                if($this->data['button']['add_access_levels']){
+                    echo "<a href='" .URLADM . "add-access-levels/index' class='btn-success'>Cadastrar</a> ";
+                }
+                if($this->data['button']['sync_pages_levels']){
+                    echo "<a href='" .URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                }
                 ?>
             </div>
         </div>
@@ -50,12 +54,25 @@
                                 class="dropdown-btn-action"><?php echo ICON_SETTINGS ?> Ações</button>
                             <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                 <?php
+                                if($this->data['button']['order_access_levels']){
                                     echo "<a href='" . URLADM . "order-access-levels/index/$id?pag=" . $this->data['pag'] . "'>". ICON_ORDER." Ordem</a>";
-                                        
-                                    echo "<a href='" . URLADM . "list-permission/index?level=$id'>". ICON_PERMISSION." Permissão</a>";
-                                    echo "<a href='" . URLADM . "view-access-levels/index/$id'>". ICON_VIEW." Visualizar</a>";
-                                    echo "<a href='" . URLADM . "edit-access-levels/index/$id'>". ICON_EDIT." Editar</a>";
-                                    echo "<a href='" . URLADM . "delete-access-levels/index/$id' onclick='return confirm(\"Tem certeza que deseja excuir este Registro?\")'>". ICON_DELETE." Apagar</a>";
+
+                                }
+                                    if($this->data['button']['list_permission']){
+                                        echo "<a href='" . URLADM . "list-permission/index?level=$id'>". ICON_PERMISSION." Permissão</a>";
+                                    }
+                                    if($this->data['button']['view_access_levels']){
+                                        echo "<a href='" . URLADM . "view-access-levels/index/$id'>". ICON_VIEW." Visualizar</a>";
+
+                                    }
+                                    if($this->data['button']['edit_access_levels']){
+                                        echo "<a href='" . URLADM . "edit-access-levels/index/$id'>". ICON_EDIT." Editar</a>";
+
+                                    }
+                                    if($this->data['button']['delete_access_levels']){
+                                        echo "<a href='" . URLADM . "delete-access-levels/index/$id' onclick='return confirm(\"Tem certeza que deseja excuir este Registro?\")'>". ICON_DELETE." Apagar</a>";
+
+                                    }
                                 ?>
                             </div>
                         </div>
