+++ adm/app/adms/Views/colors/listColors.php
@@ -14,7 +14,9 @@
             <span class="title-content">Listar Cores</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "add-colors/index' class='btn-success'>Cadastrar</a>";
+                if($this->data['button']['add_colors']){
+                    echo "<a href='" . URLADM . "add-colors/index' class='btn-success'>".ICON_ADD." Cadastrar</a>";
+                }
                 ?>
             </div>
         </div>
@@ -32,7 +34,7 @@
                     <th class="list-head-content">ID</th>
                     <th class="list-head-content">Nome</th>
                     <th class="list-head-content table-sm-none">Cor</th>
-                    <th class="list-head-content"><?php echo ICON_SETTINGS ?>-Ações</th>
+                    <th class="list-head-content"><?php echo ICON_SETTINGS ?> Ações</th>
                 </tr>
             </thead>
             <tbody class="list-body">
@@ -48,12 +50,19 @@
                         </td>
                         <td class="list-body-content">
                             <div class="dropdown-action">
-                                <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action"><?php echo ICON_SETTINGS ?>-Ações</button>
+                                <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action"><?php echo ICON_SETTINGS ?> Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php
-                                    echo "<a href='" . URLADM . "view-colors/index/$id'>".ICON_VIEW." Visualizar</a>";
-                                    echo "<a href='" . URLADM . "edit-colors/index/$id'>".ICON_EDIT." Editar</a>";
-                                    echo "<a href='" . URLADM . "delete-colors/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>".ICON_DELETE." Apagar</a>";
+                                    if($this->data['button']['view_colors']){
+                                        echo "<a href='" . URLADM . "view-colors/index/$id'>".ICON_VIEW." Visualizar</a>";
+                                    }
+                                    if($this->data['button']['edit_colors']){
+                                        echo "<a href='" . URLADM . "edit-colors/index/$id'>".ICON_EDIT." Editar</a>";
+                                    }
+                                    if($this->data['button']['delete_colors']){
+                                        echo "<a href='" . URLADM . "delete-colors/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>".ICON_DELETE." Apagar</a>";
+
+                                    }
                                     ?>
                                 </div>
                             </div>
