+++ adm/app/adms/Views/sitsUser/viewSitsUser.php
@@ -14,11 +14,23 @@
             <span class="title-content">Detalhes da Situação</span>
             <div class="top-list-right">
                 <?php
-                echo "<a href='" . URLADM . "list-sits-users/index' class='btn-info'>Listar</a> ";
-                if (!empty($this->data['viewSitUser'])) {
-                    echo "<a href='" . URLADM . "edit-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "' class='btn-warning'>Editar</a> ";
-                    echo "<a href='" . URLADM . "delete-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                if($this->data['button']['list_sits_users']){
+
+                    echo "<a href='" . URLADM . "list-sits-users/index' class='btn-info'>Listar</a> ";
                 }
+
+                if (!empty($this->data['viewSitUser'])) {
+                    if($this->data['button']['edit_sits_users']){
+
+                        echo "<a href='" . URLADM . "edit-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+                
+                    if($this->data['button']['delete_sits_users']){
+
+                        echo "<a href='" . URLADM . "delete-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
+                }
+
                 ?>
             </div>
         </div>
